refactor permission commands into one slash command
